% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{useless_multiply}
\alias{useless_multiply}
\title{Multiply Two Numbers}
\usage{
useless_multiply(x, y)
}
\arguments{
\item{x}{A numeric value. The first number.}

\item{y}{A numeric value. The second number.}
}
\value{
The product of `x` and `y`.
}
\description{
This function takes two numeric inputs and returns their product.
}
\examples{
useless_multiply(2, 6)
useless_multiply(0, 10)
}
