% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{useless_subtract}
\alias{useless_subtract}
\title{Subtract Two Numbers}
\usage{
useless_subtract(x, y)
}
\arguments{
\item{x}{A numeric value. The first number.}

\item{y}{A numeric value. The second number.}
}
\value{
The difference of `x` and `y`.
}
\description{
This function takes two numeric inputs and returns the result of subtracting `y` from `x`.
}
\examples{
useless_subtract(7, 4)
useless_subtract(10, 20)
}
